Console: Support root prompts as separate entity (WIP)
In console applications, many users have a separate color scheme for
root prompts and user prompts to better differentiate them. Here I'm
experimenting with tokenizing these prompts separately from standard
user prompts. I've chosen the `Name.Entity` token for this, as a number
of Pygments themes I've seen already list this as an alias for "Root
Prompts" in their comments.

Changes:

* Added an `rprompt` option to control the delimiters for root prompts,
  and set a default value to `#`. This is then removed from the default
  values for regular prompts.
* Implemented the `rprompt_regex` and it's helper functions, patterned
  after the same functions for `prompt_regex`
* Updated some of the default functionality relating to comments that
  assumed `#` was a default `prompt` character instead of a default
  `rprompt` character

I still need to do some more testing and refactoring of this so consider
it a work in progress.