Added specs and removed '%' from rprompt defaults
Added some specs for the root prompt lexing, and ensured they passed on
Ruby 2.6.10, 2.7.6, 3.0.4, and 3.1.2.

Removed '%' from the defaults for rprompt. I had only included it
because I didn't want an empty default when comments are explicitly
enabled, but decided against since '%' is not a common *root* prompt in
practice.